fix tracecount metric
